requirements.txt
Flask==1.0.2
requests
redis
pymongo
python-binary-memcached
Pillow
